feat: setup initial ts-node server
