+++ index.html
@@ -11,6 +11,6 @@
     <!-- <div id="root"></div>
     <script type="module" src="/src/main.tsx"></script> -->
     <div id="app"><!--app-html--></div>
-    <script type="module" src="/src/entry-client.js"></script>
+    <!-- <script type="module" src="/src/entry-client.js"></script> -->
   </body>
 </html>
