+++ package.json
@@ -4,7 +4,7 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "vite",
+    "dev": "ts-node --esm server.ts",
     "build": "npm run lint && tsc && vite build",
     "preview": "vite preview",
     "lint": "npx eslint src",
@@ -43,6 +43,7 @@
     "jsdom": "^21.1.1",
     "msw": "^1.2.1",
     "prettier": "^2.8.4",
+    "ts-node": "^10.9.1",
     "typescript": "^4.9.3",
     "vite": "^4.1.0",
     "vitest": "^0.29.3"
