+++ tsconfig.json
@@ -20,6 +20,6 @@
     "noEmit": true,
     "types": ["node", "jest", "@testing-library/jest-dom", "vitest", "vite/client", "vite-plugin-svgr/client"]
   },
-  "include": ["src"],
+  "include": ["src", "server.ts"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
